Move GUI related functions from `video.h` to `gui/common.h`
